разделяет main.js на отдельные модули, устанавливает main.js главным модулем
